requirements.txt
Babel==2.3.4
blinker==1.4
cffi==1.7.0
click==6.6
coverage==4.1
cryptography==1.4
decorator==4.0.10
ecdsa==0.13
enum34==1.1.6
Flask==0.11
Flask-Babel==0.11.1
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.12
flipflop==1.0
future==0.15.2
guess-language==0.2
hashids==1.1.0
idna==2.1
ipaddress==1.0.16
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pbr==1.10.0
py==1.4.31
pyOpenSSL==16.0.0
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
pytest==2.9.2
python-jose==1.0.0
python-openid==2.2.5
pytz==2016.4
six==1.10.0
SQLAlchemy==1.0.13
sqlalchemy-migrate==0.10.0
sqlparse==0.1.19
Tempita==0.5.2
Werkzeug==0.11.10
Whoosh==2.7.4
WTForms==2.1
